Code Delete User Function

+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/UserService.java
@@ -8,12 +8,14 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Objects;
 
 @Service
 public class UserService {
 
     @Autowired
-    private  UserRepository userRepo;
+    private final UserRepository userRepo;
 
     @Autowired
     private RoleRepository roleRepo;
@@ -31,7 +33,16 @@ public class UserService {
 
     public void save(User user) {
         boolean isUpdatingUser = (user.getId() != null);
-         encodePassword(user);
+        if (isUpdatingUser) {
+            User existingUser = userRepo.findById(user.getId()).get();
+            if (user.getPassword().isEmpty()) {
+                user.setPassword(existingUser.getPassword());
+            } else {
+                encodePassword(user);
+            }
+        } else {
+            encodePassword(user);
+        }
         userRepo.save(user);
     }
 
@@ -40,16 +51,40 @@ public class UserService {
         user.setPassword(encodedPassword);
     }
 
-    public User get(Integer id) {
-        return userRepo.findById(id).get();
+    public User isEmailUnique(Integer id, String email) {
+        User userByEmail = userRepo.getUserByEmail(email);
+        if (userByEmail == null) return null;
+
+        boolean isCreatingNew = (id == null);
+        if (isCreatingNew) {
+            return userByEmail;
+        } else {
+            if (!Objects.equals(userByEmail.getId(), id)) {
+                return userByEmail;
+            }
+        }
+
+        return null;
     }
 
-    public void delete(Integer id) {
-        userRepo.deleteById(id);
+    public User get(Integer id) throws UserNotFoundException {
+        try {
+            return userRepo.findById(id).get();
+        } catch (NoSuchElementException ex) {
+            throw new UserNotFoundException("Could not find any user with ID " + id);
+        }
     }
 
     public List<Role> listRoles() {
         return (List<Role>) roleRepo.findAll();
+    }
+
+    public void delete(Integer id) throws UserNotFoundException {
+        Long countById = userRepo.countById(id);
+        if (countById == null || countById == 0) {
+            throw new UserNotFoundException("Could not find any user with ID " + id);
+        }
+        userRepo.deleteById(id);
     }
 
 }
